+++ src/lib.rs
@@ -20,7 +20,7 @@ extern crate llvm_sys;
 use std::io;
 use std::io::Write;
 
-use libc::c_uint;
+use libc::{c_uint, uintptr_t};
 
 use llvm_sys::prelude::*;
 use llvm_sys::core::*;
@@ -214,11 +214,11 @@ fn test_ee() {
     struct TestMM;
 
     impl execution_engine::SimpleMCJITMemoryManagerImpl for TestMM {
-        fn allocate_code_section(&mut self, _size: u32, _alignment: u32, _section_id: u32, _section_name: &str) -> *mut u8 {
+        fn allocate_code_section(&mut self, _size: uintptr_t, _alignment: u32, _section_id: u32, _section_name: &str) -> *mut u8 {
             0 as *mut u8
         }
 
-        fn allocate_data_section(&mut self, _size: u32, _alignment: u32, _section_id: u32, _section_name: &str, _is_read_only: bool) -> *mut u8 {
+        fn allocate_data_section(&mut self, _size: uintptr_t, _alignment: u32, _section_id: u32, _section_name: &str, _is_read_only: bool) -> *mut u8 {
             0 as *mut u8
         }
